Fixed left-bar reposition on left show/hide.
Since hiding left also hides above-left,
the left-bar was jumping to the top on hide
and then back under above-left on show.
This makes left-bar stay at the top.

New .grid-portal style allows a div with this
style to act as a relative position root
for dialogs, snackbars, ribbons, etc.