DFNTC-13872 progress on modal
simplistic html modal, helping understand what is possible with the
scaffold API and what is hard.

this commit knows how to use the  "portal-wide" location to block
activity on mys of the page, but is unable to block activity on the
ribbon or title bar yet.

might need to explore adding either a new location type, or create the
modal oncept in the framework directly and allow a panel to be added
there while open, then removed when closed, and only one thing can ever
be there?