Implement "blocked extensions".
Suppose an extension is leaking memory,
this admin function allows you to configure a list of URLs
to block form loading so that you can attempt to
verify that the system is fine without the extension running.

Refactor kbar-results/kbar-routes to reduce broiler plate
command function that just switches routes.